# references/taoswap-api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Taoswap API",
    "version": "1.0.0",
    "description": "API documentation",
    "license": {
      "name": "BSD License"
    }
  },
  "paths": {
...
    "/subnets/": {
      "get": {
        "operationId": "subnets_retrieve",
        "description": "Subnet aggregated API.\nUses DRF serializer but keeps explicit ViewSet control.",
        "tags": [
          "subnets"
        ],
        "security": [
          {}
        ],
        "responses": {
          "200": {
            "description": "No response body"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      }
    }
  }
}